+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Migrate recipe cost to parsed values
 
+### Fixed
+
+* Filter out lower bonuses in inspect settlement
+
 ## [4.8.0] - 2025-04-09
 
 ### Added